A revamp of the strategy to encourage subclassing
In the previous implementation, each option got evaluated and passed cartain
variables from the strategy which was a litle awkward.

This became a problem as when trying to implement multiple policies I was
not able to store the policy_name in the session as it as the session
variable was not passed to the `policy` option.

The new approach is to encourage subclassing of the strategy.
This gives the developer greater control over their integration.
It also looks a lot nicer.